Refs #26167 -- Moved Query to FunctionalIndexQuery
