+++ django/db/models/sql/query.py
@@ -34,7 +34,7 @@ from django.utils.encoding import force_text
 from django.utils.functional import cached_property
 from django.utils.tree import Node
 
-__all__ = ['Query', 'RawQuery']
+__all__ = ['FunctionalIndexQuery', 'Query', 'RawQuery']
 
 
 def get_field_names_from_opts(opts):
@@ -2078,3 +2078,14 @@ class JoinPromoter:
         query.promote_joins(to_promote)
         query.demote_joins(to_demote)
         return to_demote
+
+
+class FunctionalIndexQuery:
+
+    def __init__(self, model):
+        self.model = model
+
+    def resolve_ref(self, name, *args, **kwargs):
+        if not any(f.name == name for f in self.model._meta.concrete_fields):
+            raise ValueError("Invalid reference to field '%s' on model '%s'" % (name, self.model._meta.label))
+        return Ref(name, None)
